RED_SOCIAL_PYTHON-main/requirements.txt
mysql-connector-python==9.4.0
prettytable==3.16.0
